Move task API from api app to tasks app and remove api app
- Move api_utils.py, base_views.py, base_serializers.py to tasks app
- Move task views, serializers, urls to apps/tasks/v1/
- Create apps/tasks/urls.py with proper namespacing
- Update URL namespace from api:v1:tasks:* to tasks:v1:*
- Remove apps/api/ directory entirely
- Update INSTALLED_APPS and pyproject.toml
- Use automatic package discovery in pyproject.toml
- Add missing apps/health/__init__.py
- Update all related tests and imports
