--- apps/api/__init__.py
@@ -1,1 +1,0 @@
-# This file makes api a Python package
