--- apps/api/v1/__init__.py
@@ -1,1 +1,0 @@
-# This file makes v1 a Python package
