--- apps/api/v1/tasks/__init__.py
@@ -1,3 +1,0 @@
-"""
-Task-related API endpoints package.
-"""
