+++ tests/unit/general/test_urls_basic.py
@@ -37,7 +37,7 @@ class BasicURLConfigurationTestCase(TestCase):
             import apps.dashboard.urls  # noqa: F401
 
         with contextlib.suppress(ImportError):
-            import apps.api.urls  # noqa: F401
+            import apps.tasks.urls  # noqa: F401
 
     def test_django_url_functions(self):
         """Test that Django URL functions work correctly."""
@@ -98,7 +98,7 @@ class BasicURLConfigurationTestCase(TestCase):
         # Test that we can create the same patterns as in the actual config
         patterns = [
             path("dashboard/", include("apps.dashboard.urls")),
-            path("api/", include("apps.api.urls")),
+            path("api/", include("apps.tasks.urls")),
         ]
 
         # All patterns should be valid
